+++ providers/vimeo.py
@@ -20,17 +20,35 @@ from providers import Provider
 import re
 
 class Vimeo(Provider):
-    DEFAULT = ''
+    DEFAULT = set(['hd', 'sd'])
     FORMATS = {
-        '<omit>': '640x360 H.264/AAC Stereo MP4',
+        'sd': '640x360 H.264/AAC Stereo MP4',
         'hd': '1280x720 H.264/AAC Stereo MP4',
     }
 
     def __init__(self, id, **kwargs):
         Provider.__init__(self, id, **kwargs)
 
-        self.format = kwargs.pop('format', Vimeo.DEFAULT)
-        self.debug('Vimeo', '__init__', 'format', self.format)
+        self.format = kwargs.pop('format', None)
+        self.extension = 'mp4' #Set via download_callback but needs a default
+        self._html = None
+        self._formats = None
+
+
+    @property
+    def html(self):
+        if self._html is None:
+            self._html = Provider._download(self._get_data_url()).read()
+        return self._html
+
+    @property
+    def formats(self):
+        if self._formats is None:
+            self._formats = set(['sd'])
+            match = re.search(r'<isHD>(0|1)<\/isHD>', self.html)
+            if match and match.group(1) == '1':
+                self._formats.add('hd')
+        return self._formats
 
 
     def get_title(self):
@@ -39,39 +57,48 @@ class Vimeo(Provider):
             title = match.group(1).decode('utf-8')
         else:
             title = Provider.get_title(self)
-        self.debug('Vimeo', 'get_title', 'title', title)
+        self._debug('Vimeo', 'get_title', 'title', title)
         return title
 
     def get_filename(self):
         filename = "%s.%s" % (self.get_title(), self.extension)
-        self.debug('Vimeo', 'get_filename', 'filename', filename)
+        self._debug('Vimeo', 'get_filename', 'filename', filename)
         return filename
-
-    def download_callback(self, url):
-        self.extension = re.search(r'(mp4|flv)', url.geturl()).group(1)
-        self.debug('Vimeo', 'download_callback', 'extension', self.extension)
-
-    def get_data_url(self):
-        url = 'http://vimeo.com/moogaloop/load/clip:%s' % self.id
-        self.debug('Vimeo', 'get_data_url', 'url', url)
-        return url
 
     def get_download_url(self):
         url = 'http://vimeo.com/moogaloop/play/clip:%s/%s/%s/' % (self.id, self._get_signature(), self._get_signature_expiration())
-        if self.format:
+        if self.format is None:
+            self.format = self._get_best_format()
+        if self.format != 'sd':
             url += '?q=%s' % self.format
-        self.debug('Vimeo', 'get_download_url', 'url', url)
+        self._debug('Vimeo', 'get_download_url', 'url', url)
         return url
 
+    def download_callback(self, url):
+        self.extension = re.search(r'(mp4|flv)', url.geturl()).group(1)
+        self._debug('Vimeo', 'download_callback', 'extension', self.extension)
+
+
+    def _get_best_format(self):
+        for format in Vimeo.DEFAULT:
+            if format in self.formats:
+                self._debug('Vimeo', '_get_best_format', 'format', format)
+                return format
+        raise ValueError("Could not determine the best available format. Vimeo has likely changed its page layout. Please contact the author of this script.")
+
+    def _get_data_url(self):
+        url = 'http://vimeo.com/moogaloop/load/clip:%s' % self.id
+        self._debug('Vimeo', 'get_data_url', 'url', url)
+        return url
 
     def _get_signature(self):
         match = re.search(r'<request_signature>(.+)<\/request_signature>', self.html)
         value = match.group(1) if match else None
-        self.debug('Vimeo', '_get_signature', 'signature', value)
+        self._debug('Vimeo', '_get_signature', 'signature', value)
         return value
 
     def _get_signature_expiration(self):
-        match = re.search (r'<request_signature_expires>(.+)<\/request_signature_expires>', self.html)
+        match = re.search(r'<request_signature_expires>(.+)<\/request_signature_expires>', self.html)
         value = match.group(1) if match else None
-        self.debug('Vimeo', '_get_signature_expiration', 'expiration', value)
+        self._debug('Vimeo', '_get_signature_expiration', 'expiration', value)
         return value