+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 Change Log
 ==========
 
+Version 1.2.0 *(In Development)*
+--------------------------------
+New Features:
+
+ * Ability to determine valid formats for a specific video.
+ * Automatic downloading of highest format available if none specified.
+
 Version 1.1.0 *(2010-04-04)*
 --------------------------------
 New Features: